requirements.txt
django
djangorestframework
sqlalchemy
